Remove api folder, moved variable processing
